requirements.txt
asgiref==3.11.0
Django==6.0.1
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
inflection==0.5.1
packaging==25.0
pillow==12.1.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
